{{cookiecutter.tool_identifier}}/requirements.txt
flask
mwoauth
pyyaml
requests
toolforge
